added/modified tests, made sparse_objective work for long

+++ uncurl/nolips.pyx
@@ -104,9 +104,9 @@ def sparse_objective(X, np.ndarray[DTYPE_t, ndim=2] M, np.ndarray[DTYPE_t, ndim=
     cdef double obj = 0
     # use a csc matrix, iterate through 
     X_csc = sparse.csc_matrix(X)
-    cdef int[:] indices, indptr
-    indices = X_csc.indices
-    indptr = X_csc.indptr
+    cdef long[:] indices, indptr
+    indices = X_csc.indices.astype(np.int64)
+    indptr = X_csc.indptr.astype(np.int64)
     cdef double[:] data_ = X_csc.data
     cdef double[:] mw = np.zeros(len(data_))
     cdef double[:,:] m_view = M
